moved validation.ts to usercontroller
